Root resolvers moved to separate components.
