+++ api/src/instantiation/instantiator.component.ts
@@ -1,24 +1,24 @@
 import * as _ from 'lodash';
-import { Component } from '@nestjs/common';
-import { Config } from '../config/config.component';
-import { StagesListFactory } from './stages-list-factory.component';
-import { Build } from './build';
-import { Source } from './source';
-import { JobInterface } from './job/job';
-import { CopyBeforeBuildTaskJob } from './job/copy-before-build-task.job';
-import { InterpolateBeforeBuildTaskJob } from './job/interpolate-before-build-task.job';
-import { CreateDirectoryJob } from './job/create-directory.job';
-import { ResolveReferenceJob } from './job/resolve-reference.job';
-import { DownloadSourceJob } from './job/download-source.job';
-import { ExtractSourceJob } from './job/extract-source.job';
-import { ParseDockerComposeJob } from './job/parse-docker-compose.job';
-import { PreparePortDomainsJob } from './job/prepare-port-domains.job';
-import { PrepareEnvironmentalVariablesJob } from './job/prepare-environmental-variables.job';
-import { PrepareSummaryItemsJob } from './job/prepare-summary-items.job';
-import { RunDockerComposeJob } from './job/run-docker-compose.job';
-import { GetContainerIdsJob } from './job/get-container-ids.job';
-import { ConnectContainersToNetworkJob } from './job/connect-containers-to-network-job';
-import { ProxyPortDomainsJob } from './job/proxy-port-domains.job';
+import {Component} from '@nestjs/common';
+import {Config} from '../config/config.component';
+import {StagesListFactory} from './stages-list-factory.component';
+import {Build} from './build';
+import {Source} from './source';
+import {JobInterface} from './job/job';
+import {CopyBeforeBuildTaskJob} from './job/copy-before-build-task.job';
+import {InterpolateBeforeBuildTaskJob} from './job/interpolate-before-build-task.job';
+import {CreateDirectoryJob} from './job/create-directory.job';
+import {ResolveReferenceJob} from './job/resolve-reference.job';
+import {DownloadSourceJob} from './job/download-source.job';
+import {ExtractSourceJob} from './job/extract-source.job';
+import {ParseDockerComposeJob} from './job/parse-docker-compose.job';
+import {PreparePortDomainsJob} from './job/prepare-port-domains.job';
+import {PrepareEnvironmentalVariablesJob} from './job/prepare-environmental-variables.job';
+import {PrepareSummaryItemsJob} from './job/prepare-summary-items.job';
+import {RunDockerComposeJob} from './job/run-docker-compose.job';
+import {GetContainerIdsJob} from './job/get-container-ids.job';
+import {ConnectContainersToNetworkJob} from './job/connect-containers-to-network-job';
+import {ProxyPortDomainsJob} from './job/proxy-port-domains.job';
 import {BaseLogger} from '../logger/base-logger';
 
 @Component()
